refactor(filid): extract compressPaths to utils organ (T2-D half-split)
intent-injector half-split per rfx single-responsibility-file rule.
compressPaths moved to new utils/compress-paths.ts organ; intent-injector.ts
drops from 310 to 251 lines. Facade re-exports compressPaths so src/index.ts
surface is unchanged.

- src/index.ts diff: empty (backward compat preserved)
- bundle delta: 0% (all bridge bundles unchanged)
- INTENT.md untouched
